Add Compositor::destroySurface()
Alo change how we handle the actual wayland surface destruction:
If the client decides to destroy the surface then we set it to null and
destroy the container. Setting it to null avoids accidentally
dereferencing the garbage object in the SurfaceContainer destructor.
If we are destroying the surface upon the decision of the server then we just
delete the container and it will take care of the rest